// lib/rick-and-morty-api-client/models/Character.ts
export interface Character {
    /**
     * 
     * @type {number}
     * @memberof Character
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    species?: string;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    gender?: string;
    /**
     * 
     * @type {CharacterOrigin}
     * @memberof Character
     */
    origin?: CharacterOrigin;
    /**
     * 
     * @type {CharacterLocation}
     * @memberof Character
     */
    location?: CharacterLocation;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    image?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Character
     */
    episode?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Character
     */
    url?: string;
    /**
     * 
     * @type {Date}
     * @memberof Character
     */
    created?: Date;
}